+++ nginx-nomad-alt/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Version bump for new base image
 * Move php-fpm enable to cook script
 * Setup steps for manual setup php-fpm service
+* Adjust for php-fpm/php_fpm in RC service, undo prior changes
 
 ---